frameworks: base: Systemui: Fix file manager crashing issue when open the seafile application.

+++ packages/DocumentsUI/src/com/android/documentsui/StartupMenuAdapter.java
@@ -99,25 +99,27 @@ public class StartupMenuAdapter extends BaseAdapter {
                     Cursor c = mdb.rawQuery("select * from perpo where pkname = ?",
                                             new String[] { pkgName });
                     c.moveToNext();
-                    int numbers = c.getInt(c.getColumnIndex("int"));
-                    numbers++;
-                    int number = c.getInt(c.getColumnIndex("click"));
-                    number++;
-                    ContentValues values = new ContentValues();
-                    values.put("int", numbers);
-                    values.put("click", number);
-                    mdb.update("perpo", values, "pkname = ?", new String[] { pkgName });
-                    SharedPreferences sharedPreference = mContext.getSharedPreferences("click",
-                                                                         Context.MODE_PRIVATE);
-                    SharedPreferences.Editor editor = sharedPreference.edit();
-                    String type = sharedPreference.getString("type", "sortName");
-                    int order = sharedPreference.getInt("order", 0);
-                    editor.clear();
-                    editor.putInt("isClick", 1);
-                    editor.putString("type", type);
-                    editor.putInt("order", order);
-                    editor.putInt("isSql", 1);
-                    editor.commit();
+                    if (c.moveToFirst()) {
+                        int numbers = c.getInt(c.getColumnIndex("int"));
+                        numbers++;
+                        int number = c.getInt(c.getColumnIndex("click"));
+                        number++;
+                        ContentValues values = new ContentValues();
+                        values.put("int", numbers);
+                        values.put("click", number);
+                        mdb.update("perpo", values, "pkname = ?", new String[] { pkgName });
+                        SharedPreferences sharedPreference = mContext.getSharedPreferences("click",
+                                                                             Context.MODE_PRIVATE);
+                        SharedPreferences.Editor editor = sharedPreference.edit();
+                        String type = sharedPreference.getString("type", "sortName");
+                        int order = sharedPreference.getInt("order", 0);
+                        editor.clear();
+                        editor.putInt("isClick", 1);
+                        editor.putString("type", type);
+                        editor.putInt("order", order);
+                        editor.putInt("isSql", 1);
+                        editor.commit();
+                    }
                     break;
                 case MotionEvent.BUTTON_TERTIARY:
                     break;
